fix automatic migration issues
