+++ components/Accordion/index.tsx
@@ -1,68 +1,59 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-import classnames from 'classnames';
-// @ts-expect-error TS(7016): Could not find a declaration file for module 'reac... Remove this comment to see the full error message
-import { useRef, useState } from 'react';
-import ChevronIcon from '../Icons/Chevron';
+import classnames from "classnames";
+import { PropsWithChildren, useRef, useState } from "react";
+import ChevronIcon from "../Icons/Chevron";
 
+interface Props extends PropsWithChildren {
+  initOpen: boolean;
+  header: string;
+  headerClassName?: string;
+  bodyClassName?: string;
+  tabIndex?: number;
+}
 export default function Accordion({
-  initOpen,
+  initOpen = false,
   header,
   headerClassName,
   bodyClassName,
-  tabIndex,
-  children
-}: any) {
+  tabIndex = 0,
+  children,
+}: Props) {
   const [isOpen, setIsOpen] = useState(initOpen);
-  const ref = useRef();
+  const ref = useRef<HTMLDivElement>(null);
 
   const bodyHeight = isOpen
     ? { height: ref.current?.scrollHeight }
     : { height: 0 };
 
   return (
-    // @ts-expect-error TS(17004): Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
     <>
-      {/* @ts-expect-error TS(7026): JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message */}
       <div
         role="button"
         className={classnames(
-          'flex justify-between items-center focus:outline-none p-3',
+          "flex justify-between items-center focus:outline-none p-3",
           headerClassName
         )}
         onClick={() => setIsOpen(!isOpen)}
         tabIndex={tabIndex}
       >
         {header}
-        {/* @ts-expect-error TS(7026): JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message */}
         <span className="float-right">
           {isOpen ? (
-            // @ts-expect-error TS(17004): Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
             <ChevronIcon className="-rotate-90" />
           ) : (
-            // @ts-expect-error TS(17004): Cannot use JSX unless the '--jsx' flag is provided... Remove this comment to see the full error message
             <ChevronIcon className="rotate-90" />
           )}
-        {/* @ts-expect-error TS(7026): JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message */}
         </span>
-      {/* @ts-expect-error TS(7026): JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message */}
       </div>
-      {/* @ts-expect-error TS(7026): JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message */}
       <div
         className={classnames(
-          'overflow-hidden md:overflow-x-hidden transition-height ease duration-300 border-t pt-1',
+          "overflow-hidden md:overflow-x-hidden transition-height ease duration-300 border-t pt-1",
           bodyClassName
         )}
         ref={ref}
         style={bodyHeight}
       >
         {children}
-      {/* @ts-expect-error TS(7026): JSX element implicitly has type 'any' because no i... Remove this comment to see the full error message */}
       </div>
     </>
   );
 }
-
-Accordion.defaultProps = {
-  initOpen: false,
-  tabIndex: 0,
-};
