--- components/Icons/Mail.ts
@@ -1,33 +1,0 @@
-import classnames from 'classnames';
-
-export default function MailIcon({
-  classNames
-}: any) {
-  return (
-    // @ts-expect-error TS(2304): Cannot find name 'svg'.
-    <svg
-      // @ts-expect-error TS(2304): Cannot find name 'xmlns'.
-      xmlns="http://www.w3.org/2000/svg"
-      // @ts-expect-error TS(2552): Cannot find name 'className'. Did you mean 'classN... Remove this comment to see the full error message
-      className={classnames('h-6 w-6', classNames: any)}
-      // @ts-expect-error TS(2304): Cannot find name 'fill'.
-      fill="none"
-      // @ts-expect-error TS(2304): Cannot find name 'viewBox'.
-      viewBox="0 0 24 24"
-      // @ts-expect-error TS(2304): Cannot find name 'stroke'.
-      stroke="currentColor"
-    >
-      // @ts-expect-error TS(2304): Cannot find name 'path'.
-      <path
-        // @ts-expect-error TS(2304): Cannot find name 'strokeLinecap'.
-        strokeLinecap="round"
-        // @ts-expect-error TS(2304): Cannot find name 'strokeLinejoin'.
-        strokeLinejoin="round"
-        // @ts-expect-error TS(2304): Cannot find name 'strokeWidth'.
-        strokeWidth={2}
-        // @ts-expect-error TS(2304): Cannot find name 'd'.
-        d="M3 8l7.89 5.26a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
-      />
-    </svg>
-  );
-}
