+++ components/Layout/index.tsx
@@ -1,8 +1,7 @@
-export default function Layout({
-  children
-}: any) {
+import { PropsWithChildren } from "react";
+
+export default function Layout({ children }: PropsWithChildren) {
   return (
-    // @ts-expect-error TS(2304): Cannot find name 'div'.
     <div className="container min-h-screen flex flex-col justify-center items-center mx-auto py-0 px-2 sm:px-4">
       {children}
     </div>
