+++ eslint-config/next.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // __ NEXT.JS
+  'react/react-in-jsx-scope': 0,
+};
