--- eslint-config/next.ts
@@ -1,5 +1,0 @@
-// @ts-expect-error TS(2580): Cannot find name 'module'. Do you need to install ... Remove this comment to see the full error message
-module.exports = {
-  // __ NEXT.JS
-  'react/react-in-jsx-scope': 0,
-};
