+++ eslint-config/node.js
@@ -1,4 +1,3 @@
-// @ts-expect-error TS(2580): Cannot find name 'module'. Do you need to install ... Remove this comment to see the full error message
 module.exports = {
   'node/callback-return': 'error',
   'node/exports-style': 'error',
