+++ eslint-config/react-hooks.js
@@ -1,0 +1,4 @@
+module.exports = {
+  'react-hooks/exhaustive-deps': 'error',
+  'react-hooks/rules-of-hooks': 'error',
+};
