+++ lib/stringUtils.ts
@@ -1,4 +1,6 @@
-// @ts-expect-error TS(7031): Binding element 'first' implicitly has an 'any' ty... Remove this comment to see the full error message
-export const capitalize = ([first, ...rest], lowerRest = false) =>
+export const capitalize = (
+  [first, ...rest]: string,
+  lowerRest = false
+): string =>
   first.toUpperCase() +
-  (lowerRest ? rest.join('').toLowerCase() : rest.join(''));
+  (lowerRest ? rest.join("").toLowerCase() : rest.join(""));
