+++ lib/types.ts
@@ -1,0 +1,3 @@
+export type PropsWithClassName<T> = T & {
+  className?: string;
+};
