+++ package.json
@@ -24,6 +24,8 @@
     "@babel/eslint-parser": "^7.19.1",
     "@babel/preset-react": "^7.18.6",
     "@next/eslint-plugin-next": "^13.0.1",
+    "@types/node": "18.14.2",
+    "@types/react": "18.0.28",
     "babel-eslint": "^9.0.0",
     "eslint": "^8.26.0",
     "eslint-config-airbnb": "^19.0.4",
