+++ pages/_app.tsx
@@ -1,0 +1,7 @@
+import type { AppProps } from "next/app";
+
+import "../styles/global.css";
+
+export default function App({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />;
+}
