+++ pages/api/hello.ts
@@ -1,10 +1,10 @@
-/* eslint-disable sort-keys */
+import { NextApiRequest, NextApiResponse } from "next";
 
-export default (req: any, res: any) => {
+export default function handler(req: NextApiRequest, res: NextApiResponse) {
   res.statusCode = 200;
   res.json({
-    i: 'love you',
-    wont: 'you tell',
-    me: 'your name?',
+    i: "love you",
+    wont: "you tell",
+    me: "your name?",
   });
-};
+}
