+++ sections/Footer/index.tsx
@@ -1,0 +1,13 @@
+export default function Footer() {
+  return (
+    <footer className="w-full h-24 flex justify-center items-center border-t border-solid border-gray-300">
+      <a
+        className="flex justify-center items-center text-black no-underline"
+        href="https://github.com/ignaguri"
+        rel="noopener noreferrer"
+      >
+        Created with ♥ by @ignaguri
+      </a>
+    </footer>
+  );
+}
